Move Seed tests into secretseed_test.rs
